+++ src/MonadIO/T/MonadIO.hs
@@ -30,12 +30,13 @@ import TastyPlus  ( runTestsP, runTestsReplay, runTestTree )
 import qualified MonadIO.File
 import qualified MonadIO.FPath 
 import qualified MonadIO.FStat
+import qualified MonadIO.OpenFile
 
 --------------------------------------------------------------------------------
 
 tests ∷ TestTree
 tests = testGroup "MonadIO" [ MonadIO.FStat.tests, MonadIO.FPath.tests
-                            , MonadIO.File.tests ]
+                            , MonadIO.File.tests, MonadIO.OpenFile.tests ]
 
 ----------------------------------------